+++ packages/gate_next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @folie/gate-next
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @folie/gate@0.0.11
+
 ## 0.0.2
 
 ### Patch Changes