+++ playground/backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @folie/playground-backend
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @folie/gate@0.0.11
+
 ## 0.0.4
 
 ### Patch Changes